+++ src/main/java/fiskfille/tf/common/motion/TFMotionManager.java
@@ -206,11 +206,11 @@ public class TFMotionManager {
 		}
 
 		if(driftAmount > 0.2F) {
-			player.rotationYaw += driftAmount * 2;
+			player.rotationYaw += (float) (driftAmount * 2D);
 		}
 
 		if(driftAmount < -0.2F) {
-			player.rotationYaw -= -driftAmount * 2;
+			player.rotationYaw -= (float) (-driftAmount * 2D);
 		}
 	}
